chore: relocate cursor
