+++ functions/src/helpers/cursor.ts
